// src/grpc/proto/proto/main.ts
export interface EchoRequest {
  input: string
}

export interface EchoResponse {
  output: string
}

export interface SendAnalyticsRequest {
  eventName: string
  eventValue: string
}

export interface SendAnalyticsResponse {}

export interface KeychainGetPasswordRequest {
  key: string
}

export interface KeychainGetPasswordResponse {
  value: string
  found: boolean
}

export interface KeychainSetPasswordRequest {
  key: string
  value: string
}

export interface KeychainSetPasswordResponse {}

export interface KeychainDeletePasswordRequest {
  key: string
}

export interface KeychainDeletePasswordResponse {
  deleted: boolean
}

export interface KeychainFindCredentialsRequest {}

export interface KeychainFindCredentialsResponse {
  keys: string[]
}

export interface RunPeerPluginRequest {
  pluginName: string
  args: string[]
  cwd: string
}

export interface RunPeerPluginResponse {}
// ...
export const EchoRequest = {
  encode(message: EchoRequest, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.input !== '') {
      writer.uint32(10).string(message.input)
    }
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): EchoRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseEchoRequest()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break
          }

          message.input = reader.string()
          continue
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(object: any): EchoRequest {
    return { input: isSet(object.input) ? globalThis.String(object.input) : '' }
  },

  toJSON(message: EchoRequest): unknown {
    const obj: any = {}
    if (message.input !== '') {
      obj.input = message.input
    }
    return obj
  },

  create<I extends Exact<DeepPartial<EchoRequest>, I>>(base?: I): EchoRequest {
    return EchoRequest.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<EchoRequest>, I>>(object: I): EchoRequest {
    const message = createBaseEchoRequest()
    message.input = object.input ?? ''
    return message
  },
}

function createBaseEchoResponse(): EchoResponse {
  return { output: '' }
}

export const EchoResponse = {
  encode(message: EchoResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    if (message.output !== '') {
      writer.uint32(10).string(message.output)
    }
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): EchoResponse {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseEchoResponse()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break
          }

          message.output = reader.string()
          continue
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(object: any): EchoResponse {
    return { output: isSet(object.output) ? globalThis.String(object.output) : '' }
  },

  toJSON(message: EchoResponse): unknown {
    const obj: any = {}
    if (message.output !== '') {
      obj.output = message.output
    }
    return obj
  },

  create<I extends Exact<DeepPartial<EchoResponse>, I>>(base?: I): EchoResponse {
    return EchoResponse.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<EchoResponse>, I>>(object: I): EchoResponse {
    const message = createBaseEchoResponse()
    message.output = object.output ?? ''
    return message
  },
}

function createBaseSendAnalyticsRequest(): SendAnalyticsRequest {
  return { eventName: '', eventValue: '' }
}

export const SendAnalyticsRequest = {
  encode(
    message: SendAnalyticsRequest,
    writer: _m0.Writer = _m0.Writer.create(),
  ): _m0.Writer {
    if (message.eventName !== '') {
      writer.uint32(10).string(message.eventName)
    }
    if (message.eventValue !== '') {
      writer.uint32(18).string(message.eventValue)
    }
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): SendAnalyticsRequest {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseSendAnalyticsRequest()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break
          }

          message.eventName = reader.string()
          continue
        case 2:
          if (tag !== 18) {
            break
          }

          message.eventValue = reader.string()
          continue
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(object: any): SendAnalyticsRequest {
    return {
      eventName: isSet(object.eventName) ? globalThis.String(object.eventName) : '',
      eventValue: isSet(object.eventValue) ? globalThis.String(object.eventValue) : '',
    }
  },

  toJSON(message: SendAnalyticsRequest): unknown {
    const obj: any = {}
    if (message.eventName !== '') {
      obj.eventName = message.eventName
    }
    if (message.eventValue !== '') {
      obj.eventValue = message.eventValue
    }
    return obj
  },

  create<I extends Exact<DeepPartial<SendAnalyticsRequest>, I>>(
    base?: I,
  ): SendAnalyticsRequest {
    return SendAnalyticsRequest.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<SendAnalyticsRequest>, I>>(
    object: I,
  ): SendAnalyticsRequest {
    const message = createBaseSendAnalyticsRequest()
    message.eventName = object.eventName ?? ''
    message.eventValue = object.eventValue ?? ''
    return message
  },
}

function createBaseSendAnalyticsResponse(): SendAnalyticsResponse {
  return {}
}

export const SendAnalyticsResponse = {
  encode(_: SendAnalyticsResponse, writer: _m0.Writer = _m0.Writer.create()): _m0.Writer {
    return writer
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): SendAnalyticsResponse {
    const reader = input instanceof _m0.Reader ? input : _m0.Reader.create(input)
    let end = length === undefined ? reader.len : reader.pos + length
    const message = createBaseSendAnalyticsResponse()
    while (reader.pos < end) {
      const tag = reader.uint32()
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break
      }
      reader.skipType(tag & 7)
    }
    return message
  },

  fromJSON(_: any): SendAnalyticsResponse {
    return {}
  },

  toJSON(_: SendAnalyticsResponse): unknown {
    const obj: any = {}
    return obj
  },

  create<I extends Exact<DeepPartial<SendAnalyticsResponse>, I>>(
    base?: I,
  ): SendAnalyticsResponse {
    return SendAnalyticsResponse.fromPartial(base ?? ({} as any))
  },
  fromPartial<I extends Exact<DeepPartial<SendAnalyticsResponse>, I>>(
    _: I,
  ): SendAnalyticsResponse {
    const message = createBaseSendAnalyticsResponse()
    return message
  },
}
// ...
export type CoreCLIHelperService = typeof CoreCLIHelperService
export const CoreCLIHelperService = {
  echo: {
    path: '/proto.CoreCLIHelper/Echo',
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: EchoRequest) =>
      Buffer.from(EchoRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer) => EchoRequest.decode(value),
    responseSerialize: (value: EchoResponse) =>
      Buffer.from(EchoResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer) => EchoResponse.decode(value),
  },
  sendAnalytics: {
    path: '/proto.CoreCLIHelper/SendAnalytics',
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: SendAnalyticsRequest) =>
      Buffer.from(SendAnalyticsRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer) => SendAnalyticsRequest.decode(value),
    responseSerialize: (value: SendAnalyticsResponse) =>
      Buffer.from(SendAnalyticsResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer) => SendAnalyticsResponse.decode(value),
  },
  keychainGetPassword: {
    path: '/proto.CoreCLIHelper/KeychainGetPassword',
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: KeychainGetPasswordRequest) =>
      Buffer.from(KeychainGetPasswordRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer) => KeychainGetPasswordRequest.decode(value),
    responseSerialize: (value: KeychainGetPasswordResponse) =>
      Buffer.from(KeychainGetPasswordResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer) => KeychainGetPasswordResponse.decode(value),
  },
  keychainSetPassword: {
    path: '/proto.CoreCLIHelper/KeychainSetPassword',
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: KeychainSetPasswordRequest) =>
      Buffer.from(KeychainSetPasswordRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer) => KeychainSetPasswordRequest.decode(value),
    responseSerialize: (value: KeychainSetPasswordResponse) =>
      Buffer.from(KeychainSetPasswordResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer) => KeychainSetPasswordResponse.decode(value),
  },
  keychainDeletePassword: {
    path: '/proto.CoreCLIHelper/KeychainDeletePassword',
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: KeychainDeletePasswordRequest) =>
      Buffer.from(KeychainDeletePasswordRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer) => KeychainDeletePasswordRequest.decode(value),
    responseSerialize: (value: KeychainDeletePasswordResponse) =>
      Buffer.from(KeychainDeletePasswordResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer) => KeychainDeletePasswordResponse.decode(value),
  },
  keychainFindCredentials: {
    path: '/proto.CoreCLIHelper/KeychainFindCredentials',
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: KeychainFindCredentialsRequest) =>
      Buffer.from(KeychainFindCredentialsRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer) => KeychainFindCredentialsRequest.decode(value),
    responseSerialize: (value: KeychainFindCredentialsResponse) =>
      Buffer.from(KeychainFindCredentialsResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer) => KeychainFindCredentialsResponse.decode(value),
  },
  runPeerPlugin: {
    path: '/proto.CoreCLIHelper/RunPeerPlugin',
    requestStream: false,
    responseStream: false,
    requestSerialize: (value: RunPeerPluginRequest) =>
      Buffer.from(RunPeerPluginRequest.encode(value).finish()),
    requestDeserialize: (value: Buffer) => RunPeerPluginRequest.decode(value),
    responseSerialize: (value: RunPeerPluginResponse) =>
      Buffer.from(RunPeerPluginResponse.encode(value).finish()),
    responseDeserialize: (value: Buffer) => RunPeerPluginResponse.decode(value),
  },
} as const
// ...
export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>

type KeysOfUnion<T> = T extends T ? keyof T : never
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never
    }